+++ ProtocolCtrl/DEBIAN/preinst
@@ -1,9 +1,13 @@
 #!/bin/sh
 
-onvifLogDir="/etc/aqueti/Log/"
-usrName=$(who |awk -F' ' '{print $1}')
-echo "$usrName"
-sudo chown -R "$usrName":"$usrName" "/etc/aqueti"
+onvifLogDir="/var/tmp/aqueti/Log/"
+usrName=$(who |awk -F' ' '{print $1}' | head -1)
+#echo "$usrName"
+#sudo chown -R "$usrName":"$usrName" "/etc/aqueti"
+
+sudo cp /etc/aqueti/ProtocolCtrl/ProCtrl.desktop ~/Desktop
+sudo chmod +x ~/Desktop/ProCtrl.desktop
+sudo chmod +x /etc/aqueti/ProtocolCtrl/start.sh
 
 if [ ! -d "$onvifLogDir" ]; then
 	mkdir "$onvifLogDir"
