+++ ProtocolCtrl/etc/aqueti/ProtocolCtrl/start.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+cd /etc/aqueti/ProtocolCtrl/   && python2 ProtocolCtrl2.py
